refactor(ios): extract card editor managed media text
